Block create geofence if the geofence to replace doesn't exist

+++ woosmapgeofencing/src/main/java/com/webgeoservices/woosmapgeofencing/PositionsManager.kt
@@ -1366,16 +1366,21 @@ class PositionsManager(val context: Context, private val db: WoosmapDb) {
         region.type = "isochrone"
 
         Thread {
-            this.db.regionsDAO.deleteRegionFromId(oldId)
-            this.db.regionsDAO.createRegion(region)
+            val regionDB = this.db.regionsDAO.getRegionFromId(oldId)
+            if(regionDB == null) {
+                Log.d(WoosmapSdkTag, "Region to replace not exist id = " + oldId)
+            } else {
+                this.db.regionsDAO.deleteRegionFromId(oldId)
+                this.db.regionsDAO.createRegion(region)
 
-            if (Woosmap.getInstance().regionReadyListener != null) {
-                Woosmap.getInstance().regionReadyListener.RegionReadyCallback(region)
-            }
+                if (Woosmap.getInstance().regionReadyListener != null) {
+                    Woosmap.getInstance().regionReadyListener.RegionReadyCallback(region)
+                }
 
-            val lastPosition = this.db.movingPositionsDao.getLastMovingPosition()
-            if (lastPosition != null) {
-                calculateDistanceWithRegion(lastPosition,this.db.regionsDAO.regionIsochrone)
+                val lastPosition = this.db.movingPositionsDao.getLastMovingPosition()
+                if (lastPosition != null) {
+                    calculateDistanceWithRegion(lastPosition, this.db.regionsDAO.regionIsochrone)
+                }
             }
 
         }.start()
@@ -1438,24 +1443,23 @@ class PositionsManager(val context: Context, private val db: WoosmapDb) {
         region.type = "circle"
 
         Thread {
-            this.db.regionsDAO.deleteRegionFromId(oldId)
-            val regionDB = this.db.regionsDAO.getRegionFromId(newId)
-            if(regionDB != null) {
-                Log.d(WoosmapSdkTag, "Region already exist")
+            val regionDB = this.db.regionsDAO.getRegionFromId(oldId)
+            if(regionDB == null) {
+                Log.d(WoosmapSdkTag, "Region to replace not exist id = " + oldId)
             } else {
+                this.db.regionsDAO.deleteRegionFromId(oldId)
                 this.db.regionsDAO.createRegion(region)
+                mGeofencingClient.addGeofences(geofencingRequest, GeofencePendingIntent).run {
+                    addOnSuccessListener {
+                        Log.d(WoosmapSdkTag,"onSuccess: Geofence Added...")
+                    }
+                    addOnFailureListener {
+                        val errorMessage = geofenceHelper.getErrorString(exception)
+                        Log.d(WoosmapSdkTag,"onFailure "+errorMessage)
+                    }
+                }
             }
         }.start()
-
-        mGeofencingClient.addGeofences(geofencingRequest, GeofencePendingIntent).run {
-            addOnSuccessListener {
-                Log.d(WoosmapSdkTag,"onSuccess: Geofence Added...")
-            }
-            addOnFailureListener {
-                val errorMessage = geofenceHelper.getErrorString(exception)
-                Log.d(WoosmapSdkTag,"onFailure "+errorMessage)
-            }
-        }
     }
 
 }
